String rewriter: Add length preserving rewrite
This commit adds methods to the string rewriter to transform a string
into a normalized string with the same length (but not the same content
necessarily). One of the use cases is normalizing the string before the
start index in an indexof operation. For example:

```
str.indexof(str.++("ABCD", x), y, 3) --->
str.indexof(str.++("AAAD", x), y, 3)
```

In addition to the helper methods, this commit adds the indexof rewrite
above.