Rewrite subtring length if used in replacement
In cases such as (str.replace x (str.++ ... (str.substr y i j)) z), we
can restrict j to len(x) + 1 - len(...) if j > len(x) + 1. Because if
the string to be replaced is longer than x, then it does not matter how
much longer it is, the result is always x. Thus, it is fine to only look
at the prefix of length len(x) + 1 - len(...).