Fix build issue for Clang
The build was failing for Clang as "run.cpp" includes "base/check.h"
which is a private library and in turn includes "cvc5_private_library.h".
This makes "run.cpp" also private (mostly unintentionally?) and so
we need to add the "CVC5_EXPORT" decorator to the run command to use it
outside.

Note: It is still not clear why this issue is only seen with Clang
and not other compilers/platforms.