Update the expected input interface to invoke oracles
Earlier Apporach:
While calling the executable, we used to merge all the different inputs
to the executable and wrap them inside single quotes. Say, our executable
is named isLessThan and takes two ints 1 and 2, the call to the executable
would be:
isLessThan '1 2'

Problem with the earlier approach:
Say our inputs were a string "Hello World" and an integer 2, we would pass
the following to the executable: '"Hello World" 2' (as a single parameter).
This would mean that the executable would then have to take its arg[0]
and parse it intelligently; it could not just split on " " because we
can have spaces in strings and we get ["Hello, World", 2] on naively
splitting on space and not the intended ["Hello World", 2].
This made parsing of arguments difficult in executables.

Now:
We now pass the different arguments to the executable separately.
Like:
isLessThan 1 2
anotherExecutable '"Hello World"' 2

The executables can now access the different arguments as
arg[0] (= "Hello World"), arg[1] (= 2) and so on. They need not worry about
splitting into separate parameters (as this is done while calling the executable)
and can just handle each parameter independently.